Fix "Event system cannot take large numbers" #1277707. Lua's default number type is double, not ints - and any number over 999999 was getting converted into 0.

+++ src/TLuaInterpreter.cpp
@@ -10419,7 +10419,7 @@ bool TLuaInterpreter::callEventHandler( QString & function, TEvent * pE )
     {
         if( pE->mArgumentTypeList[i] == ARGUMENT_TYPE_NUMBER )
         {
-            lua_pushnumber( L, pE->mArgumentList[i].toInt() );
+            lua_pushnumber( L, pE->mArgumentList[i].toDouble() );
         }
         else
         {
